// src/enums/Skill.enum.ts
export enum Skill {
  AIRBYTE = "AIRBYTE",
  AIRTABLE = "AIRTABLE",
  ANGULAR = "ANGULAR",
  BOOTSTRAP = "BOOTSTRAP",
  CPP = "C++",
  CSS = "CSS",
  DAGSTER = "DASGTER",
  DBT = "DBT",
  EXPRESS = "EXPRESS",
  GIT = "GIT",
  HTML = "HTML",
  JAVASCRIPT = "JAVASCRIPT",
  JEST = "JEST",
  KERAS = "KERAS",
  MATLAB = "MATLAB",
  METABASE = "METABASE",
  MONGO = "MONGO",
  MONGOOSE = "MONGOOSE",
  NODE = "NODE",
  POSTGRES = "POSTGRES",
  PYTHON = "PYTHON",
  SUPERSET = "SUPERSET",
  TYPESCRIPT = "TYPESCRIPT",
}